adds org view to profile page
