+++ app/dashboard/models.py
@@ -1317,10 +1317,12 @@ class Profile(SuperModel):
         default=False,
         help_text='If this option is chosen, the user is able to submit a faucet/ens domain registration even if they are new to github',
     )
+    organizations = ArrayField(models.CharField(max_length=200), blank=True, default=[])
+    keywords = ArrayField(models.CharField(max_length=200), blank=True, default=[])
     form_submission_records = JSONField(default=[], blank=True)
     # Sample data: https://gist.github.com/mbeacom/ee91c8b0d7083fa40d9fa065125a8d48
     max_num_issues_start_work = models.IntegerField(default=3)
-    preferred_payout_address = models.CharField(max_length=255, default='')
+    preferred_payout_address = models.CharField(max_length=255, default='', blank=True)
     max_tip_amount_usdt_per_tx = models.DecimalField(default=500, decimal_places=2, max_digits=50)
     max_tip_amount_usdt_per_week = models.DecimalField(default=1500, decimal_places=2, max_digits=50)
