+++ app/dashboard/views.py
@@ -40,6 +40,7 @@ from django.views.decorators.http import require_GET, require_POST
 
 from app.utils import ellipses, sync_profile
 from avatar.utils import get_avatar_context
+from dashboard.utils import profile_helper
 from economy.utils import convert_amount
 from gas.utils import conf_time_spread, gas_advisories, gas_history, recommend_min_gas_price_to_confirm_in_time
 from git.utils import get_auth_url, get_github_user_data, is_github_token_valid
@@ -915,61 +916,6 @@ def quickstart(request):
 class ProfileHiddenException(Exception):
     pass
 
-
-def profile_helper(handle, suppress_profile_hidden_exception=False):
-    """Define the profile helper.
-
-    Args:
-        handle (str): The profile handle.
-
-    Raises:
-        DoesNotExist: The exception is raised if a Profile isn't found matching the handle.
-            Remediation is attempted by syncing the profile data.
-        MultipleObjectsReturned: The exception is raised if multiple Profiles are found.
-            The latest Profile will be returned.
-
-    Returns:
-        dashboard.models.Profile: The Profile associated with the provided handle.
-
-    """
-    try:
-        profile = Profile.objects.get(handle__iexact=handle)
-    except Profile.DoesNotExist:
-        profile = sync_profile(handle)
-        if not profile:
-            raise Http404
-    except Profile.MultipleObjectsReturned as e:
-        # Handle edge case where multiple Profile objects exist for the same handle.
-        # We should consider setting Profile.handle to unique.
-        # TODO: Should we handle merging or removing duplicate profiles?
-        profile = Profile.objects.filter(handle__iexact=handle).latest('id')
-        logging.error(e)
-
-    if profile.hide_profile and not profile.is_org and not suppress_profile_hidden_exception:
-        raise ProfileHiddenException
-
-    return profile
-
-
-def profile_keywords_helper(handle):
-    """Define the profile keywords helper.
-
-    Args:
-        handle (str): The profile handle.
-
-    """
-    profile = profile_helper(handle, True)
-
-    keywords = []
-    for repo in profile.repos_data:
-        language = repo.get('language') if repo.get('language') else ''
-        _keywords = language.split(',')
-        for key in _keywords:
-            if key != '' and key not in keywords:
-                keywords.append(key)
-    return keywords
-
-
 def profile_keywords(request, handle):
     """Display profile keywords.
 
@@ -977,11 +923,11 @@ def profile_keywords(request, handle):
         handle (str): The profile handle.
 
     """
-    keywords = profile_keywords_helper(handle)
+    profile = profile_helper(handle, True)
 
     response = {
         'status': 200,
-        'keywords': keywords,
+        'keywords': profile.keywords,
     }
     return JsonResponse(response)
